refactor: model 폴더명 domain으로 변경
